Extract outside click HOC and helpers
