+++ src/hoc/index.ts
@@ -3,3 +3,4 @@ export * from './withResponsiveMode';
 export * from './withValidation';
 export * from './withOutsideClick';
 export * from './withFormContext';
+export * from './withClickOutsideFC';
